# Cache type ID for composite types