--- a/runtime/sema/authaccount_contracts.go
+++ b/runtime/sema/authaccount_contracts.go
@@ -71,7 +71,7 @@ var AuthAccountContractsType = func() *CompositeType {
 
 func init() {
 	// Set the container type after initializing the `AuthAccountContractsType`, to avoid initializing loop.
-	AuthAccountContractsType.ContainerType = AuthAccountType
+	AuthAccountContractsType.SetContainerType(AuthAccountType)
 }
 
 const authAccountContractsTypeAddFunctionDocString = `