--- a/runtime/sema/check_composite_declaration.go
+++ b/runtime/sema/check_composite_declaration.go
@@ -853,7 +853,7 @@ func (checker *Checker) explicitInterfaceConformances(
 ) []*InterfaceType {
 
 	var interfaceTypes []*InterfaceType
-	seenConformances := map[TypeID]bool{}
+	seenConformances := map[*InterfaceType]bool{}
 
 	for _, conformance := range declaration.Conformances {
 		convertedType := checker.ConvertType(conformance)
@@ -861,9 +861,7 @@ func (checker *Checker) explicitInterfaceConformances(
 		if interfaceType, ok := convertedType.(*InterfaceType); ok {
 			interfaceTypes = append(interfaceTypes, interfaceType)
 
-			typeID := interfaceType.ID()
-
-			if seenConformances[typeID] {
+			if seenConformances[interfaceType] {
 				checker.report(
 					&DuplicateConformanceError{
 						CompositeType: compositeType,
@@ -873,7 +871,7 @@ func (checker *Checker) explicitInterfaceConformances(
 				)
 			}
 
-			seenConformances[typeID] = true
+			seenConformances[interfaceType] = true
 
 		} else if !convertedType.IsInvalidType() {
 			checker.report(