## Description

Cache type ID for composite types, so that it doesn't have to recalculate again and again. This was showing up on the profiler.

Benchmarking shows it could gain up to 25% improvement for certain methods. e.g: declareNonEnumCompositeValue

_Note: These benchmark results are on top of #948. Improvement against the direct master-branch is even higher._

```
name                           old time/op    new time/op    delta
DeclareNonEnumCompositeValue   58.2µs ±40%    43.2µs ±29%  -25.77%  (p=0.040 n=10+5)

name                           old alloc/op   new alloc/op   delta
DeclareNonEnumCompositeValue    30.0kB ± 0%    27.3kB ± 0%   -9.01%  (p=0.000 n=9+5)

name                           old allocs/op  new allocs/op  delta
DeclareNonEnumCompositeValue       576 ± 0%       485 ± 0%  -15.80%  (p=0.000 n=10+5)
```
<!--
Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels