@turbolent 2021-05-27
Thank you for looking into optimizing this Supun!

We had tried this before, in #658, but then saw issues on Canary and reverted it in #670. Do we know now what caused the problems and can be sure we won't repeat the same issue? 

Also, like in #658, we need to make this thread-safe, as the types are potentially accessed concurrently in the host environment (program caching)