@turbolent 2021-05-27
@SupunS here is the whole thread discussing the problem and the revert: 
https://axiomzen.slack.com/archives/C01E28D03FV/p1615274743014800?thread_ts=1615273189.014700&cid=C01E28D03FV