@turbolent 2021-05-27
@SupunS We only quickly reverted it to get that release deployed, but I just want to say that it is worth it and we should still try to get this in. 

I think the fix would be to invalidate the cached ID when the container type changes. However, there might be a whole chain of container types, so if a parent's container type changes, also its children's IDs need to be updated (recursively). We currently only have a pointer from the child to the container, I think we should make the `ContainerType` field private and add a setter function (e.g. `SetContainerType`), which also informs the children (recursively) about the change, so they can update their IDs.

This isn't really a problem right now / does not exist at the moment, as only one level of containment is possible: non-contract types may be contained in a contract. However, in the future we might change this, so we shouldn't rely