@SupunS 2021-05-28
Right, that makes sense.

I think I found the issue: Seems like it's happening for nested interfaces. 
Here at [check_composite_declaration.go#L864](https://github.com/onflow/cadence/blob/master/runtime/sema/check_composite_declaration.go#L864) `interfaceType.ID()` is called and there it caches the typeID. And later on at [check_interface_declaration.go#L262](https://github.com/onflow/cadence/blob/master/runtime/sema/check_interface_declaration.go#L262), `ContainerType` of the interface is updated. 

I will add a setter for `ContainerType` and invalidate the cache upon updates, like you suggested. Thanks for the tips!