@turbolent 2021-05-29
@SupunS Nice work, thank you for finding the spot where this happens. I agree, in this case it should be OK. I'm still a bit worried that there are potentially other cases or future cases where it is not OK. Could we potential mark the read of the ID in the case where it is OK (because it is only used temporarily) as "safe if invalidated later", and other cases would be "panic if the ID changes"? This could be just a boolean on the cache which is checked by the setter. Happy to sync on details