@SupunS 2021-05-29
Yeah, that's a good safety measure.

It seems this is the only place where the ID is cached before setting the `containerType`. We can avoid that as well, by using a `map[*InterfaceType]bool{}` assuming types are singleton?

Then we can simply panic if the ID is cached always. Added this change with 2e22c7d

In the future, if we have a valid use-case where the ID is needed, then maybe we can add a method to retrieve the ID without caching. But I feel we don't need to add it at this point, so that such usages are discouraged automatically. WDYT?