@turbolent 2021-06-01
I guess when I wrote the code and used the type ID my thinking was that the conformances might be type aliases at some point, but I guess we can use `map[*InterfaceType]bool{}` for now, yes.

2e22c7d looks good!